+++ src/services/report/lib/vega/index.ts
@@ -4,6 +4,7 @@ import { join } from 'node:path';
 import { registerFont } from 'canvas';
 import { compile as handlebars } from 'handlebars';
 import type { ImageOptions } from 'jspdf';
+import { contrast } from 'chroma-js';
 import {
   cloneDeep,
   get,
@@ -12,10 +13,10 @@ import {
 } from 'lodash';
 
 import {
+  parse,
   expressionFunction,
   Locale as VegaLocale,
-  parse,
-  scheme,
+  scheme as vegaScheme,
   View,
 } from 'vega';
 import { compile, type TopLevelSpec } from 'vega-lite';
@@ -31,26 +32,12 @@ import { calcVegaFormat } from '~/models/recurrence';
 import localeFR from './locales/fr-FR.json';
 import VegaLogger from './logger';
 
-const { outDir } = config.report;
-
-registerFont('lib/vega/fonts/Roboto-light.ttf', { family: 'Roboto', weight: 'normal' });
-registerFont('lib/vega/fonts/Roboto-medium.ttf', { family: 'Roboto', weight: 'bold' });
-// Default colors of vega
-const colorScheme = scheme('tableau10') as string[];
-// Default colors of labels for default colors of Vega (https://vega.github.io/vega/docs/schemes/)
-scheme('tableau10-labels', [
-  'white', // white on blue
-  'black', // black on orange
-  'black', // black on red
-  'black', // black on cyan
-  'black', // black on green
-  'black', // black on yellow
-  'black', // black on purple
-  'black', // black on pink
-  'white', // white on brown
-  'black', // black on grey
-]);
-
+type CanvasRegisterableFont = {
+  path: string;
+  family: string;
+  weight?: string;
+  style?: string
+};
 /**
  * Params for createVegaLSpec
  */
@@ -89,6 +76,26 @@ type VegaParams = {
 };
 
 export type InputVegaParams = Omit<VegaParams, 'width' | 'height'> & { title: Title };
+
+const {
+  outDir,
+  scheme,
+  fontFamily,
+  fonts,
+} = config.report;
+
+// Register fonts in Vega
+fonts.forEach(({ path, ...font }: CanvasRegisterableFont) => {
+  registerFont(path, font);
+  logger.verbose(`[vega] Register font: [${path}] as [${font.family} ${font.weight || ''} ${font.style || ''}]`);
+});
+
+// Colors of vega (https://vega.github.io/vega/docs/schemes/)
+const colorScheme = vegaScheme(scheme) as string[];
+
+// Colors of labels for colors of Vega
+const labelScheme = `${scheme}.labels`;
+vegaScheme(labelScheme, colorScheme.map((c) => (contrast(c, 'black') > 5 ? 'black' : 'white')));
 
 /**
  * Ratio between outer and inner radius.
@@ -204,26 +211,39 @@ export const createVegaLSpec = (
       break;
   }
 
-  let colorsEntries: [string, string][] = [];
+  const colorsEntries = new Map<string, string>();
   if (colorField) {
-    const f = colorField;
-    colorsEntries = data.map((d, i): [string, string] => {
-      const key = get(d, f);
-      let color = params.colorMap.get(key);
-      if (!color) {
-        color = colorScheme[i % colorScheme.length];
-        params.colorMap.set(key, color);
+    const labelField = colorField;
+    const unusedColorsSet = new Set(colorScheme);
+
+    const labels = new Set(data.map((el): string => get(el, labelField)));
+    // eslint-disable-next-line no-restricted-syntax
+    for (const label of [...labels]) {
+      const color = params.colorMap.get(label);
+      if (color) {
+        // Use known color
+        colorsEntries.set(label, color);
+        // Remove color from unused
+        unusedColorsSet.delete(color);
+        // Remove from labels since we used it
+        labels.delete(label);
       }
-      return [key, color];
-    });
+    }
+
+    // Set leftovers with unused colors
+    const unusedColors = [...unusedColorsSet];
+    // eslint-disable-next-line no-restricted-syntax
+    for (const label of labels) {
+      colorsEntries.set(label, unusedColors.shift() || '');
+    }
   }
 
   // Getting default encoding
   let encoding: Encoding = {
     color: {
       scale: {
-        domain: colorsEntries.map(([key]) => key),
-        range: colorsEntries.map(([, color]) => color),
+        domain: Array.from(colorsEntries.keys()),
+        range: Array.from(colorsEntries.values()),
       },
     },
   };
@@ -380,7 +400,7 @@ export const createVegaLSpec = (
           legend: null,
           scale: {
             // @ts-ignore
-            scheme: 'tableau10-labels',
+            scheme: labelScheme,
           },
         },
         detail: params.color && {
@@ -430,7 +450,7 @@ export const createVegaLSpec = (
     config: {
       locale: localeFR as VegaLocale,
       customFormatTypes: true,
-      font: 'Roboto',
+      font: fontFamily,
     },
   };